Reverted name changes and Json/Xml formatter virtual class.
When the name is owned by the nodes parent std::pair will not work at all, unless the node holds a pointer to the parent causing more issues than it's worth.

Once name is held by the node the XML node can hold a name so the system for Node Formatter can be reverted back to a static linked function call from a template.

With names held by nodes the assignment operators (not normal node copy constructor) should not change the name of the node.